Rename internal api folder to src
